requirements.txt
anyio==3.5.0
asyncclick==8.0.3.2
idna==3.3
importlib-metadata==4.10.1
pkg-resources==0.0.0
python-kasa==0.4.1
pywebio==1.5.2
sniffio==1.2.0
tornado==6.1
typing-extensions==4.0.1
ua-parser==0.10.0
user-agents==2.2.0
zipp==3.7.0
